Relocate parent-related views under one folder, small fixes
