+++ frontend/src/components/App.tsx
@@ -7,10 +7,10 @@ import { Dispatch } from 'redux';
 import LoginPage from './loginPage/loginPage';
 import QRPage from './QRPage/QRPage';
 import A2hs from './A2hs';
-import ParentRedirectView from './ParentRedirectView';
-import RegistrationView from './RegistrationView';
+import ParentRedirectView from './ParentRegistration/MainView';
+import RegistrationView from './ParentRegistration/RegistrationView';
 import ProtectedRoute from './ProtectedRoute';
-import LogoutView from './LogoutPage';
+import LogoutView from './ParentRegistration/LogoutView';
 import { userTypes, userActions } from '../types/userTypes';
 import { authTypes, authActions } from '../types/authTypes';
 import { AppState } from '../reducers';
